TS: IPScan/Search - still missing some VF changes
