+++ src/components/IPScan/Search/index.tsx
@@ -1,0 +1,422 @@
+import React, { FormEvent, PureComponent, RefObject } from 'react';
+import { connect } from 'react-redux';
+import { createSelector } from 'reselect';
+import { Helmet } from 'react-helmet-async';
+
+import { createJob, goToCustomLocation } from 'actions/creators';
+import loadable from 'higherOrder/loadable';
+
+import AdvancedOptions from 'components/IPScan/AdvancedOptions';
+import Redirect from 'components/generic/Redirect';
+import Link from 'components/generic/Link';
+import Callout from 'components/SimpleCommonComponents/Callout';
+import Button from 'components/SimpleCommonComponents/Button';
+
+import id from 'utils/cheap-unique-id';
+import blockEvent from 'utils/block-event';
+import { askNotificationPermission } from 'utils/browser-notifications';
+import { schemaProcessDataPageSection } from 'schema_org/processors';
+
+import SequenceInput, {
+  SequenceChecks,
+  SequenceInputHandle,
+} from './SequenceInput';
+import example from './example.fasta';
+
+import cssBinder from 'styles/cssBinder';
+
+import blocks from 'styles/blocks.css';
+import local from './style.css';
+import searchPageCss from 'pages/Search/style.css';
+import buttonCSS from 'components/SimpleCommonComponents/Button/style.css';
+import fonts from 'EBI-Icon-fonts/fonts.css';
+
+const css = cssBinder(searchPageCss, local, blocks, buttonCSS, fonts);
+
+export const MAX_NUMBER_OF_SEQUENCES = 100;
+
+const SchemaOrgData = loadable({
+  loader: () => import(/* webpackChunkName: "schemaOrg" */ 'schema_org'),
+  loading: () => null,
+});
+
+const checkedSelectorFor = (name: string) => `input[name="${name}"]:checked`;
+
+const getCheckedApplications = (form: HTMLFormElement) =>
+  Array.from(
+    form.querySelectorAll<HTMLInputElement>(checkedSelectorFor('appl')),
+    (input: HTMLInputElement) => input.value,
+  );
+
+const isXChecked = (x: string) => (form: HTMLElement) =>
+  !!form.querySelector(checkedSelectorFor(x));
+
+const isStayChecked = isXChecked('stay');
+
+type Props = {
+  createJob: typeof createJob;
+  goToCustomLocation: typeof goToCustomLocation;
+  ipScan?: ParsedURLServer;
+  value?: string | null;
+  main?: string;
+  search?: Record<string, string | boolean>;
+};
+
+type State = {
+  title: string | null;
+  initialAdvancedOptions: Record<string, string | boolean> | null;
+  submittedJob: string | null;
+  sequenceChecks: SequenceChecks | null;
+  dragging: boolean;
+};
+
+export class IPScanSearch extends PureComponent<Props, State> {
+  _formRef: RefObject<HTMLFormElement>;
+  _editorRef: RefObject<SequenceInputHandle>;
+  sequenceToSet?: string | null;
+
+  constructor(props: Props) {
+    super(props);
+
+    let initialAdvancedOptions = null;
+
+    if (props.search) {
+      initialAdvancedOptions = props.search;
+    }
+    this.state = {
+      title: null,
+      initialAdvancedOptions,
+      submittedJob: null,
+      sequenceChecks: null,
+      dragging: false,
+    };
+
+    this._formRef = React.createRef();
+    this._editorRef = React.createRef();
+    if (this.props.value) this.sequenceToSet = this.props.value;
+  }
+
+  componentDidUpdate() {
+    if (this.sequenceToSet && this._editorRef.current) {
+      this._handleReset(this.sequenceToSet);
+      this.sequenceToSet = null;
+    }
+  }
+
+  changeTitle = () => {
+    const newTitle = (
+      this._formRef.current?.querySelector<HTMLInputElement>(
+        'input[name="local-title"]',
+      )?.value || ''
+    ).trim();
+    this.setState({
+      title: newTitle,
+    });
+  };
+
+  _handleReset = (text?: string | unknown) => {
+    if (this._formRef.current && typeof text !== 'string') {
+      const inputsToReset = Array.from(
+        this._formRef.current.querySelectorAll<HTMLInputElement>(
+          'input[name]:not([name="stay"])',
+        ),
+      );
+      for (const input of inputsToReset) {
+        input.checked = !!input.dataset.defaultchecked;
+      }
+    }
+
+    this.setState(
+      {
+        dragging: false,
+        title: null,
+        submittedJob: null,
+      },
+      () => this._editorRef.current?.focusEditor(),
+    );
+    this._editorRef.current?.reset(typeof text === 'string' ? text : undefined);
+  };
+
+  _handleSubmit = (event: FormEvent) => {
+    event.preventDefault();
+    if (!this._formRef.current) return;
+    const editorContent = this._editorRef.current?.getContent();
+    if (!editorContent) return;
+    const localID = id(`internal-${Date.now()}`);
+    this.props.createJob({
+      metadata: {
+        localID,
+        group: this.state.title,
+        type: 'InterProScan',
+      },
+      data: {
+        input: editorContent,
+        applications: getCheckedApplications(this._formRef.current),
+      },
+    });
+    // Request browser notification
+    askNotificationPermission();
+
+    if (isStayChecked(this._formRef.current)) {
+      this._handleReset();
+      this.setState({ submittedJob: localID });
+    } else {
+      this.props.goToCustomLocation({
+        description: {
+          main: { key: 'result' },
+          result: { type: 'InterProScan' /* , accession: localID */ },
+        },
+      });
+    }
+  };
+
+  _handleFile = (file: File) => {
+    const fr = new FileReader();
+    fr.onload = () => {
+      this._handleReset(fr.result);
+      this.setState({
+        title: file.name,
+      });
+    };
+    fr.readAsText(file);
+  };
+
+  _loadExample = () => this._handleReset(example);
+
+  _cleanUp = () => this._editorRef.current?.cleanUp();
+
+  _handleDroppedFiles = blockEvent(({ dataTransfer }: DragEvent) => {
+    const file = dataTransfer?.files?.[0];
+    if (file) this._handleFile(file);
+  });
+
+  _handleDragging = blockEvent(() => this.setState({ dragging: true }));
+
+  _handleUndragging = blockEvent(() => this.setState({ dragging: false }));
+
+  _handleFileChange = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+    const file = target.files?.[0];
+    if (file) {
+      this._handleFile(file);
+      target.value = '';
+    }
+  };
+
+  render() {
+    // If we had a value, we used it in the constructor+componentDidUpdate
+    // But we don't want to have it in the url, so remove the value from it.
+    if (this.props.value)
+      return (
+        <Redirect
+          to={{
+            description: {
+              main: { key: 'search' },
+              search: { type: 'sequence' },
+            },
+          }}
+        />
+      );
+    const { dragging } = this.state;
+    return (
+      <section className={css('vf-stack', 'vf-stack--400')}>
+        <form
+          onSubmit={this._handleSubmit}
+          onDrop={this._handleDroppedFiles}
+          onDrag={this._handleDragging}
+          onDragStart={this._handleDragging}
+          onDragEnd={this._handleUndragging}
+          onDragOver={this._handleDragging}
+          onDragEnter={this._handleDragging}
+          onDragExit={this._handleUndragging}
+          onDragLeave={this._handleUndragging}
+          className={css('search-form', { dragging })}
+          ref={this._formRef}
+        >
+          <div>
+            <div
+              className={css(
+                'simple-box',
+                'border',
+                'margin-bottom-none',
+                'ipscan-block',
+              )}
+            >
+              <header>Search your sequence, in FASTA format</header>
+              <SchemaOrgData
+                data={{
+                  name: 'Search By Sequence',
+                  description: 'Search for InterPro matches in your sequences',
+                }}
+                processData={schemaProcessDataPageSection}
+              />
+              <div className={css('vf-stack', 'vf-stack--200')}>
+                {this.props.main === 'search' && this.state.submittedJob && (
+                  <Callout type="info">
+                    Your search job(
+                    <span className={css('mono')}>
+                      {this.state.submittedJob}
+                    </span>
+                    ) has been submitted. You can check its state in the{' '}
+                    <Link
+                      to={{
+                        description: {
+                          main: { key: 'result' },
+                          result: { type: 'InterProScan' },
+                        },
+                      }}
+                    >
+                      Results page
+                    </Link>
+                  </Callout>
+                )}
+                {this.props.main === 'search' && (
+                  <div className={css('description')}>
+                    <Helmet>
+                      <title>InterProScan</title>
+                    </Helmet>
+                    <p>
+                      This form enables you to submit sequences to the
+                      InterProScan web service for scanning against the InterPro
+                      protein signature databases.
+                      <br />
+                      Please note that you can submit up to{' '}
+                      {MAX_NUMBER_OF_SEQUENCES} sequences at a time.
+                      Alternatively, you can{' '}
+                      <Link
+                        to={{
+                          description: { other: ['about', 'interproscan'] },
+                        }}
+                      >
+                        download InterProScan
+                      </Link>{' '}
+                      to scan your sequences locally.
+                    </p>
+                  </div>
+                )}
+                <SequenceInput
+                  value={this.props.value}
+                  ref={this._editorRef}
+                  onChecksChange={(tests) => {
+                    this.setState({ sequenceChecks: tests });
+                  }}
+                />
+              </div>
+
+              <div className={css('row')}>
+                <div className={css('columns')}>
+                  <div className={css('button-group', 'line-with-buttons')}>
+                    <label
+                      className={`vf-button ${css(
+                        'vf-button',
+                        'vf-button--secondary',
+                        'vf-button--sm',
+                        'user-select-none',
+                      )}`}
+                    >
+                      <span
+                        className={css(
+                          'icon',
+                          'icon-common',
+                          'icon-folder-open',
+                        )}
+                      />
+                      Choose file
+                      <input
+                        type="file"
+                        onChange={this._handleFileChange}
+                        hidden
+                      />
+                    </label>
+                    <Button
+                      type="secondary"
+                      className={css('user-select-none')}
+                      onClick={this._loadExample}
+                    >
+                      Example protein sequence
+                    </Button>
+                    <Button
+                      className={css({
+                        hidden:
+                          this.state.sequenceChecks?.valid ||
+                          !this.state.sequenceChecks?.hasText,
+                      })}
+                      onClick={this._cleanUp}
+                      borderColor="var(--colors-alert-main)"
+                      backgroundColor="var(--colors-alert-main)"
+                    >
+                      Automatic FASTA clean up
+                    </Button>
+                  </div>
+                </div>
+              </div>
+
+              <AdvancedOptions
+                title={this.state.title}
+                changeTitle={this.changeTitle}
+                initialOptions={this.state.initialAdvancedOptions}
+              />
+
+              <div className={css('row')}>
+                <div
+                  className={css(
+                    'large-8',
+                    'columns',
+                    'stacked-for-small',
+                    'button-group',
+                    'margin-bottom-none',
+                    'line-with-buttons',
+                  )}
+                >
+                  <Button
+                    submit
+                    className={css({
+                      disabled: !this.state.sequenceChecks?.valid,
+                    })}
+                    disabled={
+                      this.state.sequenceChecks?.tooShort ||
+                      !this.state.sequenceChecks?.valid
+                    }
+                    value="Search"
+                  >
+                    Search
+                  </Button>
+                  <Button type="secondary" onClick={this._handleReset}>
+                    Clear
+                  </Button>
+                </div>
+                <div
+                  className={css(
+                    'large-4',
+                    'columns',
+                    'show-for-medium',
+                    'search-adv',
+                  )}
+                >
+                  <span>Powered by InterProScan</span>
+                </div>
+              </div>
+            </div>
+          </div>
+          <div className={css('dragging-overlay')}>Drop your file here</div>
+        </form>
+      </section>
+    );
+  }
+}
+
+const mapStateToProps = createSelector(
+  (state: GlobalState) => state.settings.ipScan,
+  (state: GlobalState) => state.customLocation.description,
+  (state: GlobalState) => state.customLocation.search,
+  (ipScan, desc, search) => ({
+    ipScan,
+    value: desc.search.value,
+    main: desc.main.key,
+    search,
+  }),
+);
+
+export default connect(mapStateToProps, {
+  createJob,
+  goToCustomLocation,
+})(IPScanSearch);
